move palette, display to graphics
