+++ packages/graphics/src/index.ts
@@ -2,3 +2,5 @@ export * from "./image"
 export * from "./text"
 export * from "./dotmatrix"
 export * from "./image_spi"
+export * from "./palette"
+export * from "./display"
